ai_cv_service_backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.12
slowapi==0.1.9
celery==5.4.0
redis==5.2.1
python-json-logger==2.0.7
email-validator==2.2.0
httpx==0.28.1
google-api-python-client==2.154.0
google-auth-oauthlib==1.2.1
pypdf==5.4.0
python-docx==1.1.2
